refactor(core): consolidate IME composition guard into utils/ime
The IME composition guard — the check that a keydown is committing or
cancelling a Korean/Japanese/Chinese composition and must not be treated as a
command — was implemented three different ways: a shared isImeKeyEvent
predicate whose canonical definition lived inside hooks/useFocusTrap.ts, several
hand-inlined isComposing/keyCode checks, and one ref-tracking variant. That
scattering is why the same bug keeps being rediscovered one component at a time.

Move the canonical predicate to a discoverable utils/ime.ts with the single
authoritative doc comment (why the composing keydown precedes compositionend,
and why the legacy keyCode fallback is still load-bearing). Name that fallback
value as IME_PROCESSING_KEY_CODE rather than a bare 229. Re-export from the
utils barrel and migrate the inline sites to it.

isImeKeyEvent is not part of the public API (not in src/index.ts) and every
consumer now imports it from utils, so no back-compat shim is needed.

No behavior change — the predicate logic is identical.
