+++ packages/core/src/hooks/index.ts
@@ -11,12 +11,16 @@
  * SYNC: When modified, update this header
  */
 
-export {
-  hasActiveFocusTrapEscape,
-  isImeKeyEvent,
-  useFocusTrap,
-} from './useFocusTrap';
+export {hasActiveFocusTrapEscape, useFocusTrap} from './useFocusTrap';
 export type {UseFocusTrapOptions, UseFocusTrapReturn} from './useFocusTrap';
+
+/**
+ * @deprecated Import from `@astryxdesign/core/utils` instead — this is a pure
+ * predicate, not a hook, and the `hooks` barrel is a `'use client'` boundary.
+ * Re-exported here for one release so consumers can move; will be removed in
+ * an upcoming major.
+ */
+export {isImeKeyEvent} from '../utils/ime';
 
 export {useAnnounce} from './useAnnounce';
 export type {AnnounceFn, AnnouncePoliteness} from './useAnnounce';
