+++ packages/core/src/utils/index.ts
@@ -107,3 +107,5 @@ export {rtlStyles} from './rtlStyles';
 // The shared focus ring, exported for @astryxdesign/lab — same reason
 // rtlStyles is: a style that must be identical across packages, not copied.
 export {focusOutlineStyles, focusOutlineProps} from './focusOutline.stylex';
+
+export {isImeKeyEvent} from './ime';
